Move page files into nested folder
